refactor: Auth out of `MonitorRepository` and into `ApiRespository` base class
