+++ app/src/repos/monitor-repo.ts
@@ -1,4 +1,5 @@
 import type { MonitorSummary, MonitorIdentity, Monitor, MonitorInformation } from '@/types/monitor'
+import { ApiRepository } from './api-repo'
 
 export interface MonitorRepoInterface {
   getMonitorInfos(): Promise<Array<MonitorInformation>>
@@ -6,10 +7,6 @@ export interface MonitorRepoInterface {
   addMonitor(monitor: MonitorSummary): Promise<Monitor>
   updateMonitor(monitor: MonitorIdentity): Promise<Monitor>
   deleteMonitor(monitor: MonitorIdentity): Promise<void>
-}
-
-type ApiResponse = {
-  data: any
 }
 
 type MonitorResp = {
@@ -23,13 +20,9 @@ type MonitorList = {
   }
 }
 
-export class MonitorRepository implements MonitorRepoInterface {
-  // TODO: Put API URL in env.
-  private readonly baseUrl = 'http://127.0.0.1:8000'
-  private readonly getAuthToken: () => string
-
+export class MonitorRepository extends ApiRepository implements MonitorRepoInterface {
   constructor(getAuthToken: () => string) {
-    this.getAuthToken = getAuthToken
+    super(getAuthToken)
   }
 
   async getMonitorInfos(): Promise<Array<MonitorInformation>> {
@@ -66,40 +59,5 @@ export class MonitorRepository implements MonitorRepoInterface {
     })
 
     return (resp as MonitorResp).data
-  }
-
-  private async sendRequest(
-    route: string,
-    method: string,
-    body: Record<string, any> | undefined = undefined
-  ): Promise<ApiResponse | undefined> {
-    const opts: Record<string, any> = {
-      method: method,
-      headers: {
-        Authorization: `Bearer ${this.getAuthToken()}`
-      }
-    }
-    if (body !== undefined) {
-      opts.headers.Accept = 'application/json'
-      opts.headers['Content-Type'] = 'application/json'
-      opts.body = JSON.stringify(body)
-    }
-
-    let response: Response | null = null
-    try {
-      response = await fetch(this.baseUrl + route, opts)
-    } catch (e) {
-      console.error('Failed to connect to the CronMon API:', e)
-      throw new Error('Failed to connect to the CronMon API.')
-    }
-
-    if (response.ok) {
-      const length = Number(response.headers.get('Content-Length'))
-      return length > 0 ? await response.json() : undefined
-    } else {
-      // We always want the JSON out of errors to get the description.
-      const json = await response.json()
-      throw new Error(json.error.description)
-    }
   }
 }
